+++ src/clients/rewriter.ts
@@ -1,20 +1,4 @@
 import { AppGraphQLClient, InstanceOptions, IOContext } from '@vtex/api'
-import { path } from 'ramda'
-
-export interface RouteIndexFiles {
-  lastChangeDate: string
-  routeIndexFiles: RouteIndexFileEntry[]
-}
-
-export interface RouteIndexFileEntry {
-  fileName: string
-  fileSize: string
-}
-
-export interface RouteIndexEntry {
-  id: string
-  lastChangeDate: string
-}
 
 export interface RedirectInput {
   id: string
@@ -38,6 +22,11 @@ export enum RedirectTypes {
   TEMPORARY = 'TEMPORARY',
 }
 
+export interface ExportResponse {
+  routes: Redirect[]
+  next: string
+}
+
 export class Rewriter extends AppGraphQLClient {
   constructor(context: IOContext, options: InstanceOptions) {
     super('vtex.rewriter@1.x', context, {
@@ -48,80 +37,36 @@ export class Rewriter extends AppGraphQLClient {
     })
   }
 
-  public routesIndexFiles = (): Promise<RouteIndexFiles> =>
+  public exportRedirects = (next?: string): Promise<ExportResponse> =>
     this.graphql
-      .query<string[], {}>(
+      .query<{ redirect: { listRedirects: ExportResponse } }, { next: string }>(
         {
           query: `
-      query RoutesIndexFiles {
+      query ListRedirects($next: String) {
         redirect {
-          indexFiles {
-            lastChangeDate
-            routeIndexFiles {
-              fileName
-              fileSize
+          listRedirects(next: $next) {
+            next
+            routes {
+              from
+              to
+              type
+              endDate
             }
           }
         }
       }
       `,
-          variables: {},
-        },
-        {
-          metric: 'rewriter-get-redirects-index-files',
-        }
-      )
-      .then(path(['data', 'redirect', 'indexFiles'])) as Promise<RouteIndexFiles>
-
-  public routesIndex = (fileName: string): Promise<RouteIndexEntry[]> =>
-    this.graphql
-      .query<string[], { fileName: string }>(
-        {
-          query: `
-      query RoutesIndex($fileName: String!) {
-        redirect {
-          index(fileName: $fileName) {
-            id
-            lastChangeDate
-          }
-        }
-      }
-      `,
-          variables: { fileName },
-        },
-        {
-          metric: 'rewriter-get-redirects-index',
-        }
-      )
-      .then(path(['data', 'redirect', 'index'])) as Promise<RouteIndexEntry[]>
-
-  public exportRedirects = (from: number, to: number): Promise<Redirect[]> =>
-    this.graphql
-      .query<Redirect[], { from: number; to: number }>(
-        {
-          query: `
-      query ListRedirects($from: Int!, $to: Int!) {
-        redirect {
-          list(from: $from, to: $to) {
-            from
-            to
-            type
-            endDate
-          }
-        }
-      }
-      `,
-          variables: { from, to },
+          variables: { next },
         },
         {
           metric: 'rewriter-get-redirects',
         }
       )
-      .then(path(['data', 'redirect', 'list'])) as Promise<Redirect[]>
+      .then(res => res.data?.redirect?.listRedirects) as Promise<ExportResponse>
 
   public importRedirects = (routes: RedirectInput[]): Promise<boolean> =>
     this.graphql
-      .mutate<boolean, { routes: RedirectInput[] }>(
+      .mutate<{ redirect: { saveMany: boolean } }, { routes: RedirectInput[] }>(
         {
           mutate: `
       mutation SaveMany($routes: [RedirectInput!]!) {
@@ -136,11 +81,11 @@ export class Rewriter extends AppGraphQLClient {
           metric: 'rewriter-import-redirects',
         }
       )
-      .then(path(['data', 'redirect', 'saveMany'])) as Promise<boolean>
+      .then(res => res.data?.redirect?.saveMany) as Promise<boolean>
 
   public deleteRedirects = (paths: string[]): Promise<boolean> =>
     this.graphql
-      .mutate<boolean, { paths: string[] }>(
+      .mutate<{ redirect: { deleteMany: boolean } }, { paths: string[] }>(
         {
           mutate: `
       mutation DeleteMany($paths: [String!]!) {
@@ -155,5 +100,5 @@ export class Rewriter extends AppGraphQLClient {
           metric: 'rewriter-delete-redirects',
         }
       )
-      .then(path(['data', 'redirect', 'deleteMany'])) as Promise<boolean>
+      .then(res => res.data?.redirect?.deleteMany) as Promise<boolean>
 }