+++ src/lib/error/ErrorKinds.ts
@@ -2,4 +2,5 @@ export const enum ErrorKinds {
   REQUEST_ERROR = 'RequestError',
   GENERIC_ERROR = 'GenericError',
   TELEMETRY_REPORTER_ERROR = 'TelemetryReporterError',
+  DEPRECATION_CHECK_ERROR = 'DeprecationCheckError',
 }