+++ src/modules/rewriter/import.ts
@@ -1,6 +1,6 @@
 import { createHash } from 'crypto'
 import { readFile, readJson, remove, writeFile } from 'fs-extra'
-import { compose, concat, difference, isEmpty, length, map, pluck, reduce, prop } from 'ramda'
+import { difference, isEmpty, length, map, pluck } from 'ramda'
 import { createInterface } from 'readline'
 import { Parser } from 'json2csv'
 
@@ -102,17 +102,6 @@ let retryCount = 0
 export default async (csvPath: string, options: any) => {
   const reset = options ? options.r || options.reset : undefined
   let indexedRoutes
-  if (reset) {
-    const indexFiles = await rewriter.routesIndexFiles().then(prop('routeIndexFiles'))
-    const indexFileNames = pluck('fileName', indexFiles) || []
-
-    const arr = []
-    for (const name of indexFileNames) {
-      arr.push(rewriter.routesIndex(name))
-    }
-
-    indexedRoutes = compose<any, any, any>(pluck('id'), reduce(concat, []))(arr)
-  }
 
   let importedRoutes
   try {