+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.92.2] - 2020-03-23
+### Fixed
+- Use new list api for exporting redirects.
+- Deprecation checking.
+- `PayloadTooLarge` error in telemetry.
+
 ## [2.92.1] - 2020-03-17
 ### Changed
 - Update release notes URL.